Fixed login issue

+++ src/middlewares/deleteComments.js
@@ -5,12 +5,19 @@ const tweetRepository = new TweetRepository();
 const commentService = new CommentService();
 
 export const deletComments = async (req, res, next) => {
-    const tweet = await tweetRepository.get(req.params.id);
+    try {
+        const tweet = await tweetRepository.get(req.params.id);
 
-    let comments = tweet.comments;
-    comments.forEach(async (comment) => {
-        await commentService.delete(comment);
-    });
+        if (tweet) {
+            let comments = tweet.comments;
+            comments.forEach(async (comment) => {
+                await commentService.delete(comment);
+            });
+        }
 
-    next();
+        next();
+    }
+    catch (error) {
+        throw error;
+    }
 }
